+++ packages/omnidoc/src/components/MDXContent.examples.tsx
@@ -1,6 +1,0 @@
-import React from 'react'
-import { MDXContent } from 'omnidoc/components'
-
-export function Basic() {
-  return <MDXContent value="# Hello World" />
-}