+++ packages/mdx-plugins/CHANGELOG.md
@@ -1,4 +1,4 @@
-# @omnidoc/mdx-plugins
+# @renoun/mdx-plugins
 
 ## 0.1.1
 
@@ -10,4 +10,4 @@
 
 ### Minor Changes
 
-- c017f16: Adds `@omnidoc/mdx-plugins` package that includes pre-configured and custom `remark` and `rehype` plugins.
+- c017f16: Adds `@renoun/mdx-plugins` package that includes pre-configured and custom `remark` and `rehype` plugins.